feat(shortcuts): enhance text-expansion shortcuts functionality
- Implemented UTF-8 decoding for triggers and expansions in shortcuts, allowing for proper handling of Unicode characters.
- Updated `Settings` structure to include a vector for text-expansion shortcuts, enabling persistence and management through the settings UI.
- Enhanced the settings UI with a dedicated page for managing shortcuts, allowing users to add, edit, and delete text-expansion entries.
- Integrated shortcut management into the input method engines, ensuring that shortcuts are applied during text input.
- Added functionality to clear and sync shortcuts with the input controller, improving user experience and efficiency.
