+++ platforms/linux/common/ffi_handle.h
@@ -7,10 +7,56 @@
 
 #include <cstdint>
 #include <string>
+#include <vector>
 
 #include "funput.h" // generated by cbindgen; on the include path via CMake
 
 namespace funput {
+
+// Decode a UTF-8 string into Unicode scalars (the inverse of appendUtf8). Invalid
+// bytes are skipped. Used to marshal gõ tắt strings to the FFI's uint32_t arrays.
+inline std::vector<uint32_t> decodeUtf8(const std::string &s) {
+    std::vector<uint32_t> out;
+    size_t i = 0;
+    const size_t n = s.size();
+    while (i < n) {
+        const auto b0 = static_cast<unsigned char>(s[i]);
+        uint32_t cp = 0;
+        size_t extra = 0;
+        if (b0 < 0x80) {
+            cp = b0;
+        } else if ((b0 & 0xE0) == 0xC0) {
+            cp = b0 & 0x1F;
+            extra = 1;
+        } else if ((b0 & 0xF0) == 0xE0) {
+            cp = b0 & 0x0F;
+            extra = 2;
+        } else if ((b0 & 0xF8) == 0xF0) {
+            cp = b0 & 0x07;
+            extra = 3;
+        } else {
+            ++i; // invalid lead byte — skip
+            continue;
+        }
+        if (i + extra >= n) break; // truncated sequence
+        bool ok = true;
+        for (size_t k = 1; k <= extra; ++k) {
+            const auto bk = static_cast<unsigned char>(s[i + k]);
+            if ((bk & 0xC0) != 0x80) {
+                ok = false;
+                break;
+            }
+            cp = (cp << 6) | (bk & 0x3F);
+        }
+        if (ok) {
+            out.push_back(cp);
+            i += extra + 1;
+        } else {
+            ++i; // malformed continuation — skip the lead byte
+        }
+    }
+    return out;
+}
 
 // Append one Unicode scalar to a UTF-8 string.
 inline void appendUtf8(std::string &out, uint32_t cp) {
@@ -55,6 +101,15 @@ public:
     void setSmartRestore(bool on) { funput_set_smart_restore(engine_, on); }
     void setEagerRestore(bool on) { funput_set_eager_restore(engine_, on); }
     void clear() { funput_clear(engine_); }
+
+    // Text-expansion shortcuts (gõ tắt). Replace the whole table by clearing then
+    // re-adding each entry (the engine is a runtime mirror of settings.json).
+    void clearShortcuts() { funput_clear_shortcuts(engine_); }
+    void addShortcut(const std::string &trigger, const std::string &expansion) {
+        const std::vector<uint32_t> t = decodeUtf8(trigger);
+        const std::vector<uint32_t> e = decodeUtf8(expansion);
+        funput_add_shortcut(engine_, t.data(), t.size(), e.data(), e.size());
+    }
 
     FunputResult process(uint32_t codepoint) { return funput_process_char(engine_, codepoint); }
     FunputResult backspace() { return funput_backspace(engine_); }
