+++ platforms/linux/common/settings.h
@@ -9,6 +9,7 @@
 
 #include <cstdint>
 #include <string>
+#include <utility>
 #include <vector>
 
 namespace funput {
@@ -27,6 +28,9 @@ struct Settings {
     // App identifiers (fcitx5 program() / WM_CLASS) that default to English on
     // focus. Owned by the Settings UI; the addon only reads them for matching.
     std::vector<std::string> excludedAppIds;
+    // Text-expansion shortcuts (gõ tắt): (trigger, expansion) pairs. Owned by the
+    // Settings UI; the addon only reads them and pushes them into the engine.
+    std::vector<std::pair<std::string, std::string>> shortcuts;
 
     // Whether `program` (fcitx5 InputContext::program()) is on the exclusion list.
     bool isExcluded(const std::string &program) const;
