+++ platforms/linux/fcitx5/src/funput_engine.cpp
@@ -61,6 +61,10 @@ void FunputEngine::applySettings() {
     handle_.setEnabled(effectiveEnabled_);
     handle_.setSmartRestore(settings_.smartRestore);
     handle_.setEagerRestore(settings_.eagerRestore);
+    handle_.clearShortcuts();
+    for (const auto &[trigger, expansion] : settings_.shortcuts) {
+        handle_.addShortcut(trigger, expansion);
+    }
     handle_.clear();
 }
